pyproject.toml
[project]
name = "selenium-parallel-framework"
version = "0.1.0"
package-mode = false
description = "Principal-Grade Selenium Parallel Execution Framework"
authors = [
    {name = "Antigravity",email = "antigravity@google.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "selenium (>=4.0.0)",
    "pytest (>=7.0.0)",
    "pytest-xdist (>=3.3.0)",
    "pydantic-settings (>=2.0.0)",
    "allure-pytest (>=2.10.0)"
]

[dependency-groups]
dev = [
    "ruff (>=0.15.4,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "black (>=26.1.0,<27.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
